This is an initial check-in of the RandomAccess code I received from David Koester late May 2004. It contains code developed by Bob Lucas, Jacqueline Chame, and possibly others. RandomAccess.c was created by merging files core_single_cpu.c utility.c onecpu.c was created by merging files star_single_cpu.c ../single_cpu.c The files need work (removing multiple #include's, adding casts, etc.)

+++ RandomAccess/onecpu.c
@@ -1,11 +1,22 @@
-/* -*- mode: C; tab-width: 2; indent-tabs-mode: nil; fill-column: 79; coding: iso-latin-1-unix -*- */
+/* -*- mode: C; tab-width: 2; indent-tabs-mode: nil; -*-
+ *
+ * See RandomAccess.h for a comprehensive description of this test and
+ * its goals.
+ *
+ * This file contains the interface for the star single cpu RandomAccess test.
+ * The test runs on all cpus in the MPI universe, but there is no communication
+ * between cpus during the process (each cpu runs its own version of the
+ * single_cpu test).  The final result is the average of the entire system.
+ *
+ * This test uses the computational core found in core_single_cpu.c
+ */
 
 #include <hpcc.h>
-
 #include "RandomAccess.h"
 
 int
-StarRandomAccess(HPCC_Params *params) {
+StarRandomAccess(HPCC_Params *params)
+{
   int myRank, commSize;
   int rv, errCount, failure = 0, failureAll = 0;
   double minGUPs, avgGUPs, maxGUPs, localGUPs;
@@ -40,9 +51,25 @@ StarRandomAccess(HPCC_Params *params) {
 
   return 0;
 }
+/* -*- mode: C; tab-width: 2; indent-tabs-mode: nil; -*- 
+ *
+ * See RandomAccess.h for a comprehensive description of this test and
+ * its goals.
+ *
+ * This file contains the interface for the single cpu RandomAccess test.  The
+ * test is only run on a single (random) node in the MPI universe, with all
+ * other CPUs stuck (in theory, idle) in an MPI_Bcast waiting for the selected
+ * CPU to finish the RandomAccess test.
+ *
+ * This test uses the computational core found in core_single_cpu.c
+ */
+
+#include <hpcc.h>
+#include "RandomAccess.h"
 
 int
-SingleRandomAccess(HPCC_Params *params) {
+SingleRandomAccess(HPCC_Params *params)
+{
   int myRank, commSize;
   int rv, errCount, rank, failure = 0;
   double localGUPs;
@@ -59,7 +86,7 @@ SingleRandomAccess(HPCC_Params *params) {
   scl *= commSize;
 
   /* select a node at random, but not node 0 */
-  for (rank = 0; ; rank = (int)(scl * rand())) {
+  for (rank = 0; ; rank = scl * rand()) {
     if (rank > 0 && rank < commSize) break;
   }
